Tweaks to running tests
