+++ src/client/testing/testController/common/resultsHelper.ts
@@ -7,6 +7,7 @@ import { TestCase } from './testCase';
 import { TestCollection } from './testCollection';
 import { TestFile } from './testFile';
 import { TestFolder } from './testFolder';
+import { getTestCaseNodes } from './testItemUtilities';
 import { PythonTestData } from './types';
 
 type SupportedTestItemType = TestFolder | TestFile | TestCollection | TestCase;
@@ -87,24 +88,6 @@ function getSafeInt(value: string, defaultValue = 0): number {
     return num;
 }
 
-async function getTestCaseNodes(
-    testNode: TestItem<PythonTestData>,
-    collection: TestItem<TestCase>[] = [],
-): Promise<TestItem<TestCase>[]> {
-    if (testNode.data instanceof TestCase) {
-        collection.push(testNode as TestItem<TestCase>);
-    }
-    const nodes = Array.from(testNode.children.values());
-    for (const node of nodes) {
-        if (node.data instanceof TestCase) {
-            collection.push(node);
-        } else {
-            await getTestCaseNodes(node, collection);
-        }
-    }
-    return Promise.resolve(collection);
-}
-
 export async function updateResultFromJunitXml(
     outputXmlFile: string,
     testNode: TestItem<SupportedTestItemType>,
@@ -128,8 +111,9 @@ export async function updateResultFromJunitXml(
 
         testCaseNodes.forEach((node) => {
             const result = junitSuite.testcase.find((t) => {
-                const id = `${t.$.file.replace(/\\/g, '/')}::${t.$.name}`;
-                return id === node.data.raw.id || node.data.raw.id.endsWith(id);
+                const idResult = `${t.$.classname}.${t.$.name}`;
+                const idNode = node.data.runId;
+                return idResult === idNode || idNode.endsWith(idResult);
             });
             if (result) {
                 if (result.error) {
